pyproject.toml
[project]
name = "chaksu"
version = "0.1.0"
description = "A brief description of your project."  # It's good practice to add a description
authors = [
    {name = "Pramath Haritz", email = "pramath.haritz@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "numpy (>=1.24.2,<1.25.0)",
    "scikit-learn (>=1.2.2,<1.3.0)",
    "transformers (>=4.27.4,<4.28.0)",
    "tqdm (>=4.64.1,<4.65.0)",
    "pandas (>=1.5.3,<1.6.0)",
    "pillow (>=9.4.0,<9.5.0)",
    "kornia (>=0.6.12,<0.7.0)",
    "ultralytics (>=8.3.69,<9.0.0)",
    "segmentation-models-pytorch (>=0.3.0,<0.4.0)",
    "flair @ git+https://github.com/jusiro/FLAIR.git",
    "googledrivedownloader (>=1.1.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
